add copy image feature for community profile chart
